==> 03_clustering_annotation/5_annotation_from_subcluster_CD45pos.Rmd <==
---
title: "Annotation (by cell type) from subclusters"
author: "Laura Jiménez Gracia"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    toc: yes
    toc_float: yes
    number_sections: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE,
                      fig.align = "center")
```

# Introduction

## Objective
In this Rmarkdown document, we are going to merge cell type identities annotated by independent cell-lineage analysis into a single one.

# Pre-processing
## Libraries
```{r warning = FALSE, message = FALSE}
library(tidyverse)
library(Seurat)
library(dittoSeq)
library(DT)
```

## Parameters
Here we will define and load objects and functions that will be used throughout the document.
```{r}
# Paths
path_project_metadata <- here::here("01_cellranger_mapping/data/POST_12_metadata.csv")
path_samples_metadata <- here::here("01_cellranger_mapping/data/POST_12_metadata_samples_all.csv")
path_r_objects <- here::here("03_clustering_annotation/results/R_objects")
path_r_tables <- here::here("03_clustering_annotation/results/tables")
path_r_figs <- here::here("03_clustering_annotation/results/figs")

# Functions
source(here::here("bin/utils.R"))

# Parameters
confounder_variables <- c("library_name", "sample_id", "genotype", "treatment")
confounder_names <- c("Library", "Sample", "Genotype", "Treatment")

metadata_variables_confounder <- c("nCount_RNA", "nFeature_RNA", "gem_id", 
                                   "pct_mt", "pct_rb", "Phase",
                                   "library_name", "sample_id", "genotype", "treatment")

# Color palette
set.seed(1234) # for reproducibility
color_palette <- Polychrome::createPalette(60, c("#fc6060", "#74f774", "#7c7cfc"))
names(color_palette) <- NULL
#Polychrome::swatch(color_palette)
```

# Load data
The data used in this Rmarkdown document comes from **POST_12** dataset.
```{r}
# Load all metadata information
metadata <- read.csv(path_project_metadata)

# Merging metadata_all tables
metadata <- metadata %>% filter(type == "cDNA" & fraction == "CD45+")

print("Libraries/Samples metadata")
DT::datatable(metadata, options = list(scrollX = TRUE))

# Load Seurat object
seurat_obj <- readRDS(paste0(path_r_objects, "/post12_cd45pos_cells1_clustering_annotation.rds"))
Idents(seurat_obj) <- "celltypes_1"
seurat_obj
```


## Integrate cell-types identity
```{r}
seurat_obj$celltypes_1 <- as.character(seurat_obj$celltypes_1)
seurat_obj$celltypes_2 <- as.character(seurat_obj$celltypes_1)


celltypes_1_list <- c("Monocytes/Macrophages", "DCs", "T cells", "Proliferative cells")
celltypes_1_filenames <- c("monomacro", "DC", "Tcells", "proliferativecells")

# Load & add cell-type annotations
for (i in 1:length(celltypes_1_list)) {
  celltype <- celltypes_1_list[i]
  filename <- celltypes_1_filenames[i]
  file_path <- paste0(path_r_objects, "/post12_cd45pos_", filename, "_annotations.rds")
  # loading annotations
  annot <- readRDS(file_path)
  # adding cell_types labels
  seurat_obj$celltypes_2[seurat_obj$celltypes_1 == celltype] <- as.character(annot$celltypes_2)
}

seurat_obj
table(seurat_obj$celltypes_1, seurat_obj$sample_id)
table(seurat_obj$celltypes_2, seurat_obj$sample_id)
```


## Remove cells
```{r}
seurat_obj <- seurat_obj[, !(seurat_obj$celltypes_2 %in% c("Non-immune cells (CD45-)",
                                                           "Low quality cells",
                                                           "Doublets Macro-T cells", 
                                                           "Doublets DC-T cells", 
                                                           "Doublets B-T cells", 
                                                           "Doublets APC-T cells"
                                                           ))]
seurat_obj
table(seurat_obj$celltypes_1, seurat_obj$sample_id)
table(seurat_obj$celltypes_2, seurat_obj$sample_id)
```


## Modify metadata
```{r}
# Rename to be collective
seurat_obj$celltypes_1[seurat_obj$celltypes_1 == "T cells"] <- "T/NK cells"

seurat_obj$celltypes_1[seurat_obj$celltypes_2 == "Proliferative Macrophages"] <- "Monocytes/Macrophages"
seurat_obj$celltypes_2[seurat_obj$celltypes_2 == "Proliferative LC"] <- "Langerhans cells"
seurat_obj$celltypes_1[seurat_obj$celltypes_2 == "Langerhans cells"] <- "Langerhans cells"

seurat_obj$celltypes_2[seurat_obj$celltypes_2 == "Mature DC"] <- "mregDCs"
seurat_obj$celltypes_2[seurat_obj$celltypes_2 == "Monocyte-derived DC"] <- "mo-DCs"
seurat_obj$celltypes_2[seurat_obj$celltypes_2 == "Proliferative cDC1"] <- "cDC1"
seurat_obj$celltypes_1[seurat_obj$celltypes_2 == "mo-DCs"] <- "DCs"
seurat_obj$celltypes_1[seurat_obj$celltypes_2 == "cDC1"] <- "DCs"

seurat_obj$celltypes_2[seurat_obj$celltypes_2 == "Proliferative DETCs"] <- "Dendritic epidermal T cells"
seurat_obj$celltypes_1[seurat_obj$celltypes_2 == "Dendritic epidermal T cells"] <- "Dendritic epidermal T cells"

seurat_obj$celltypes_1[seurat_obj$celltypes_2 == "Proliferative CD4+ T cells"] <- "T/NK cells"
seurat_obj$celltypes_1[seurat_obj$celltypes_2 == "Proliferative DγδT cells"] <- "T/NK cells"


# Reorder cells1
cell_lineages_list <- c(
                       "Monocytes/Macrophages",
                       "Langerhans cells",
                       "DCs",
                       "T/NK cells",
                       "Dendritic epidermal T cells",
                       "Neutrophils",
                       "Mast cells"
                       )

seurat_obj$celltypes_1 <- as.factor(seurat_obj$celltypes_1)
seurat_obj$celltypes_1 <- factor(x = seurat_obj$celltypes_1,
                                     levels = cell_lineages_list)


# Reorder cells 2
cell_types_list <- c(
                      "Monocytes", 
                      "Mature Monocytes",
                      "Selenop+ M2-like Macrophages",
                      "Resident M2-like Macrophages",
                      "Slamf9+ Macrophages",
                      "M1-like Macrophages",
                      "MHCII+Ccr2+ Macrophages",
                      "Proliferative Macrophages",
                      "Langerhans cells",
                      "cDC1",
                      "cDC2",
                      "Tissue-resident cDC2",
                      "mregDCs",
                      "mo-DCs",
                      "pDC",
                      "Conventional CD4+ T cells",
                      "Foxp3+ Tregs",
                      "Proliferative CD4+ T cells",
                      "Cytotoxic CD8+ T cells",
                      "DoubleNegative T cells",
                      "NK cells",
                      "ILC2",
                      "Vγ4 T cells",
                      "Vγ6 T cells",
                      "Proliferative DγδT cells",
                      "Dendritic epidermal T cells",
                      "Neutrophils",
                      "Mast cells"
                                     )

seurat_obj$celltypes_2 <- as.factor(seurat_obj$celltypes_2)
seurat_obj$celltypes_2 <- factor(x = seurat_obj$celltypes_2,
                                     levels = cell_types_list)
seurat_obj
table(seurat_obj$celltypes_1, seurat_obj$sample_id)
table(seurat_obj$celltypes_2, seurat_obj$sample_id)
```


# Explore data annotation
## Data visualization
### Cell types 1

```{r fig.width=14, fig.height=10}
# Clustering overview
cd45pos_gg_umap_clustering_celltypes_1 <- DimPlot(object = seurat_obj,
        group.by = "celltypes_1",
        label = FALSE,
        cols = color_palette
        ) +
  guides(color = guide_legend(override.aes = list(size = 5), ncol = 1))
cd45pos_gg_umap_clustering_celltypes_1

ggsave(filename = paste0(path_r_figs, "/post12_cd45pos_cells1_clustering_annotation_umap.png"),
       plot = cd45pos_gg_umap_clustering_celltypes_1)
```

Now, we will also explore lineage clusters across sample
```{r fig.width=25, fig.height=6}
DimPlot(
  object = seurat_obj,
  group.by = "celltypes_1",
  split.by = "sample_id",
  label = FALSE,
  cols = color_palette
  )  +
    theme(legend.position = "bottom") +
  guides(color = guide_legend(override.aes = list(size = 5), ncol = 5))
```

### Cell types 2

```{r fig.width=14, fig.height=10}
# Clustering overview
cd45pos_gg_umap_clustering_celltypes_2 <- DimPlot(object = seurat_obj,
        group.by = "celltypes_2",
        label = FALSE,
        cols = color_palette,
        raster = FALSE
        ) +
  guides(color = guide_legend(override.aes = list(size = 5), ncol = 1))
cd45pos_gg_umap_clustering_celltypes_2

ggsave(filename = paste0(path_r_figs, "/post12_cd45pos_cells2_clustering_annotation_umap.png"),
       plot = cd45pos_gg_umap_clustering_celltypes_2)
```

Now, we will also explore lineage clusters across sample
```{r fig.width=25, fig.height=6}
DimPlot(
  object = seurat_obj,
  group.by = "celltypes_2",
  split.by = "sample_id",
  label = FALSE,
  cols = color_palette
  ) + 
  theme(legend.position = "bottom") +
  guides(color = guide_legend(override.aes = list(size = 5), ncol = 6))
```


## Cell proportions
### Cell types 1
```{r fig.width=14, fig.height=5}
cd45pos_gg_cell_num <- dittoBarPlot(seurat_obj, 
                            "celltypes_1",
                            group.by = "sample_id",
                            scale = "count",
                            retain.factor.levels = TRUE,
                            legend.show = FALSE,
                            color.panel = color_palette)

cd45pos_gg_cell_prop <- dittoBarPlot(seurat_obj, 
                             "celltypes_1",
                             group.by = "sample_id",
                             scale = "percent",
                            retain.factor.levels = TRUE,
                             color.panel = color_palette)

cd45pos_gg_cell_num + cd45pos_gg_cell_prop
```

### Cell types 2
```{r fig.width=14, fig.height=6}
cd45pos_gg_cell_num <- dittoBarPlot(seurat_obj, 
                            "celltypes_2",
                            group.by = "sample_id",
                            scale = "count",
                            retain.factor.levels = TRUE,
                            legend.show = FALSE,
                            color.panel = color_palette)

cd45pos_gg_cell_prop <- dittoBarPlot(seurat_obj, 
                             "celltypes_2",
                             group.by = "sample_id",
                             scale = "percent",
                            retain.factor.levels = TRUE,
                             color.panel = color_palette) +
  guides(color = guide_legend(override.aes = list(size = 4), ncol = 2))

cd45pos_gg_cell_num + cd45pos_gg_cell_prop
```


## Save data
```{r}
# Save metadata
saveRDS(seurat_obj@meta.data, paste0(path_r_objects, "/post12_cd45pos_alldata_cleaned_final_metadata.rds"))
openxlsx::write.xlsx(seurat_obj@meta.data, paste0(path_r_tables, "/post12_cd45pos_alldata_cleaned_final_metadata.xlsx"))
write.csv(seurat_obj@meta.data, paste0(path_r_tables, "/post12_cd45pos_alldata_cleaned_final_metadata.csv"))


# Save annotated Seurat object
saveRDS(seurat_obj, paste0(path_r_objects, "/post12_cd45pos_alldata_cleaned_final_clustering_annotation.rds"))
#seurat_obj <- readRDS(paste0(path_r_objects, "/post12_cd45pos_alldata_cleaned_final_clustering_annotation.rds"))
```


# Session Info
```{r}
sessionInfo()
```
